### What problem does this PR solve?

Issue Number: None

Related PR: #60583

Problem Summary: Merge the latest master branch while preserving the separated OLAP and external-table file-cache controls. Propagate both session variables into TQueryOptions, replace remaining legacy cache switches in external scan and OLAP preload paths, and avoid inferring scanned table types from TQueryType when creating query-limit holders.

### Release note

Fix table-specific file-cache switches so they take effect consistently across FE and BE paths.

### Check List (For Author)

- Test: Focused FE/BE unit tests added but not run locally per request; clang-format and diff checks completed
- Behavior changed: Yes, the OLAP and external-table file-cache session variables now control all reviewed paths
- Does this need documentation: No

+++ be/src/runtime/cache/result_node.h
@@ -93,7 +93,7 @@ private:
 private:
     int64_t _partition_key;
     PCacheValue* _cache_value = nullptr;
-    size_t _data_size;
+    int64_t _data_size;
     CacheStat _cache_stat;
 };
 
@@ -175,7 +175,7 @@ private:
     UniqueId _sql_key;
     ResultNode* _prev = nullptr;
     ResultNode* _next = nullptr;
-    size_t _data_size;
+    int64_t _data_size;
     PartitionRowBatchList _partition_list;
     PartitionRowBatchMap _partition_map;
 };
